+++ crates/hyperswitch_domain_models/Cargo.toml
@@ -14,17 +14,25 @@ payouts = ["api_models/payouts"]
 
 [dependencies]
 # First party deps
-api_models = { version = "0.1.0", path = "../api_models" }
+api_models = { version = "0.1.0", path = "../api_models", features = ["errors"] }
 common_enums = { version = "0.1.0", path = "../common_enums" }
 common_utils = { version = "0.1.0", path = "../common_utils" }
 masking = { version = "0.1.0", path = "../masking" }
 diesel_models = { version = "0.1.0", path = "../diesel_models", features = ["kv_store"] }
+cards = {version = "0.1.0", path = "../cards"}
+router_derive = {version = "0.1.0", path = "../router_derive"}
 
 # Third party deps
+actix-web = "4.5.1"
 async-trait = "0.1.79"
 error-stack = "0.4.1"
 http = "0.2.12"
+mime = "0.3.17"
 serde = { version = "1.0.197", features = ["derive"] }
 serde_json = "1.0.115"
+serde_with = "3.7.0"
 thiserror = "1.0.58"
 time = { version = "0.3.35", features = ["serde", "serde-well-known", "std"] }
+url = { version = "2.5.0", features = ["serde"] }
+utoipa = { version = "4.2.0", features = ["preserve_order", "preserve_path_order", "time"] }
+
