+++ crates/hyperswitch_domain_models/src/lib.rs
@@ -1,10 +1,12 @@
 pub mod errors;
 pub mod mandates;
 pub mod payment_address;
+pub mod payment_method_data;
 pub mod payments;
 #[cfg(feature = "payouts")]
 pub mod payouts;
 pub mod router_data;
+pub mod router_request_types;
 
 #[cfg(not(feature = "payouts"))]
 pub trait PayoutAttemptInterface {}
