+++ crates/router/src/core/payouts/retry.rs
@@ -81,7 +81,7 @@ pub async fn do_gsm_multiple_connector_actions(
             }
             api_models::gsm::GsmDecision::Requeue => {
                 Err(report!(errors::ApiErrorResponse::NotImplemented {
-                    message: errors::api_error_response::NotImplementedMessage::Reason(
+                    message: errors::NotImplementedMessage::Reason(
                         "Requeue not implemented".to_string(),
                     ),
                 }))?
@@ -145,7 +145,7 @@ pub async fn do_gsm_single_connector_actions(
             }
             api_models::gsm::GsmDecision::Requeue => {
                 Err(report!(errors::ApiErrorResponse::NotImplemented {
-                    message: errors::api_error_response::NotImplementedMessage::Reason(
+                    message: errors::NotImplementedMessage::Reason(
                         "Requeue not implemented".to_string(),
                     ),
                 }))?
