+++ crates/router/src/types/api/files.rs
@@ -1,5 +1,6 @@
 use api_models::enums::FileUploadProvider;
 use masking::{Deserialize, Serialize};
+use serde_with::serde_as;
 
 use super::ConnectorCommon;
 use crate::{
@@ -47,12 +48,13 @@ impl ForeignTryFrom<&types::Connector> for FileUploadProvider {
     }
 }
 
+#[serde_as]
 #[derive(Debug, Clone, serde::Serialize)]
 pub struct CreateFileRequest {
     pub file: Vec<u8>,
     pub file_name: Option<String>,
     pub file_size: i32,
-    #[serde(serialize_with = "crate::utils::custom_serde::display_serialize")]
+    #[serde_as(as = "serde_with::DisplayFromStr")]
     pub file_type: mime::Mime,
     pub purpose: FilePurpose,
     pub dispute_id: Option<String>,
