+++ crates/router/src/utils.rs
@@ -1,7 +1,6 @@
 #[cfg(feature = "olap")]
 pub mod connector_onboarding;
 pub mod currency;
-pub mod custom_serde;
 pub mod db_utils;
 pub mod ext_traits;
 #[cfg(feature = "kv_store")]
