+++ crates/router/src/utils/connector_onboarding.rs
@@ -4,7 +4,7 @@ use error_stack::ResultExt;
 use super::errors::StorageErrorExt;
 use crate::{
     consts,
-    core::errors::{api_error_response::NotImplementedMessage, ApiErrorResponse, RouterResult},
+    core::errors::{ApiErrorResponse, NotImplementedMessage, RouterResult},
     routes::{app::settings, AppState},
     types::{self, api::enums},
 };
